+++ test/uk/gov/hmrc/preferencesadminfrontend/services/UploadServiceSpec.scala
@@ -29,6 +29,7 @@ import play.api.test.Helpers.*
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.preferencesadminfrontend.connectors.ChannelPreferencesConnector
 import uk.gov.hmrc.preferencesadminfrontend.services.model.*
+import uk.gov.hmrc.preferencesadminfrontend.services.model.csv.CsvData
 
 import java.nio.file.{ Files, Path }
 import scala.concurrent.ExecutionContext.Implicits.global
